web/requirements.txt
Django==4.2.7
django-dotenv
django-storages
django-redis
djangorestframework
django-cors-headers

boto3
flower
gunicorn
uvicorn[standard]
psycopg2-binary
requests
celery
redis
Pillow
whitenoise
channels[daphne]
channels_redis